refactor: move issue type
