+++ dashboard/src/pages/TreeDetails/Tabs/WrapperTableWithLogSheet.tsx
@@ -8,8 +8,8 @@ import { Sheet } from '@/components/Sheet';
 
 import { LogOrJsonSheetContent } from '@/components/Sheet/LogOrJsonSheetContent';
 import type { TNavigationLogActions } from '@/components/Sheet/WrapperSheetContent';
-import type { TIssue } from '@/types/general';
 import { cn } from '@/lib/utils';
+import type { TIssue } from '@/types/issues';
 
 interface TableWithLogSheetProps {
   currentLog?: number;
